+++ src/main/resources/static/styles/js.js
@@ -98,79 +98,6 @@ $(function() {
 });
 
 
-document.getElementById("submitButton").addEventListener('click', function(event) {
-
-function applyQuery(event) {
-    event.preventDefault();
-    const colorRadio = document.querySelector('input[name="colorRadio"]:checked').value;
-    const formData = new FormData(document.querySelector(".Section-1"));
-
-    console.log('Submitting request...');
-
-    fetch("/applyQuery", {
-        method: "POST",
-        body: formData.toString(),
-    })
-    .then(response => {
-        console.log('Received response:', response);
-
-        return response.json();
-    })
-    .then(data => {
-        console.log('Received data:', data); // Log the received data
-
-        // Create the table header
-        const table = document.createElement("table");
-        const thead = document.createElement("thead");
-        const headerRow = document.createElement("tr");
-
-        // Assuming your data has four columns:
-        const headers = ["Name", "Average Temperature", "Minimum Temperature", "Maximum Temperature"];
-        headers.forEach(header => {
-            const th = document.createElement("th");
-            th.textContent = header;
-            headerRow.appendChild(th);
-        });
-
-        thead.appendChild(headerRow);
-        table.appendChild(thead);
-
-        // Create table body
-        const tbody = document.createElement("tbody");
-
-        // Iterate through the retrieved data and create table rows
-        data.forEach(rowData => {
-            const row = document.createElement("tr");
-
-            // Assuming each rowData has four values in an array
-            const cells = rowData;
-            cells.forEach(cellData => {
-                const td = document.createElement("td");
-                td.textContent = cellData;
-                row.appendChild(td);
-            });
-
-            tbody.appendChild(row);
-        });
-
-        table.appendChild(tbody);
-
-        // Create a new div to hold the table
-        const tableContainer = document.createElement("div");
-
-        // Append the table to the new div
-        tableContainer.appendChild(table);
-
-        // Clear existing content and append the new div to the main container
-        const container = document.querySelector(".Section-2"); // Select a suitable container
-        container.innerHTML = "";
-        container.appendChild(tableContainer);
-    })
-}
-
-
-})
-
