+++ src/main/resources/templates/Lv2-Subtask-B.html
@@ -218,7 +218,7 @@
                             </label>     
                     </div>
 
-                    <button class = "button" id = "submitButton" type = "submit">
+                    <button class = "button" type = "submit">
                         <div class = "inside-button">Apply</div>
                     </button>
                 </div>
